Coursera Project Getting and Cleaning Data
This repo contains the results from the coursera project in Getting and
Cleaning Data. Inside the repo you will find 1. An independent tidy data
set with the average of each variable for each activity and each
subject. 2.a code book that describes the variables, the data created in
the run_analysis.R script. 3. The run_analysis.R script that performed
the analysis. 4. A README.md file that explains how all of the script
worked using which tools.